+++ frontend/super-frontend/source/src/javascripts/application.js
@@ -4,6 +4,7 @@ import RailsUjs from "@rails/ujs";
 import { Application, Controller as StimulusController } from "stimulus";
 
 import apply_template_controller from "./apply_template_controller";
+import batch from "./batch_controller";
 import clean_filter_param_controller from "./clean_filter_param_controller";
 import clean_filter_params_controller from "./clean_filter_params_controller";
 import click_outside_to_close_controller from "./click_outside_to_close_controller";
@@ -15,6 +16,7 @@ import toggle_pending_destruction_controller from "./toggle_pending_destruction_
 
 let StimulusApplication = Application.start();
 StimulusApplication.register("apply-template", apply_template_controller);
+StimulusApplication.register("batch", batch);
 StimulusApplication.register(
   "clean-filter-param",
   clean_filter_param_controller